views: Use `display: none;` to make hidden popups unclickable.
For now, we use the opacity attribute to hide the newsletter popups
in the left-top corner (for example, we use `opacity: 0;` to make
these popups invisible).

This leads to, however, somewhat confusing UI, because these popups
remain clickable even if they are in the "hidden" state.

Instead, we should use `display: none;` to make these popups invisible
so that users do not accidentally click them.