docs: Fix the default value of the `flush_at_shutdown` parameter.
For now, the documentation says that `flush_at_shutdown` gets turned
on by default for file buffers.

However, as far as I can tell, this is just plain wrong. In fact, for
a persistent buffer like buf_file, the parameter is turned *off* by
default (we can confirm this in plugin/output.rb).

This patch fixes this issue by updating the buffer section manual.