docs: Migrate "filter-modify-apache.txt" to v1.0
This patch savages the old article on how we can use filters to
transform event streams. In order to get it work with Fluentd v1.0,
the texts (including the code examples) are thoroughly rewritten.

Although most of the content is already covered by the plugin manuals,
I think a good number of users will still find this article helpful.