docs: Add a new article on integrating Fluentd with Minio.
This patch is my attempt to salvage an old, orphaned patch which
was submitted by @koolhead17 on Jun 28, 2016.

I revised the most of the content of the article to be compatible
with Fluentd v1.0, and confirmed it works using an actual Minio
server (the tested version is "2018-01-02T23:07:00Z").

Also, I added detailed annotations to the configuration example and
some screenshot image, which will make the article more friendly to
new users.