docs: Document what occurs if the retry count exceeds the limit.
In the current implementation, all the data in the output queue will
be discarded if it fails to write out a chunk within the retry limit.
Since this event necessarily entails a (possibly huge) data loss, we
definitely need to document the behaviour in a clear way.

This patch adds the documentation to the buffer-plugin-overview page.