docs: Fix the Prometheus article to support td-agent users.
This trivial patch makes the article compatible with td-agent.

Presumably, there are many users who does not even realise `fluentd`
executable has been installed in their system (this is not uncommon
since, in the first place, td-agent is supposed to abstract it away).

This patch should be helpful to these users.