docs: Explain how to set up TLS/SSL encryption in out_forward.
In e0253bf, I've updated the `in_forward` manual to include the user
instructions for setting up TLS/SSL encryption.

This is the counter-part patch, which adds an explanation on how to
connect to an encryption-enabled server to the `out_forward` manual.

With this patch applied, the task of setting up an end-to-end
encrypted forwarding topology should become (relatively) easier.