docs: Update the version of prometheus to v1
This patch updates the manual to use the newer versions (v1.0.0 or later)
of fluent-plugin-prometheus.

In the process of creating this patch, I've actually confirmed that the
instructions of the article can be followed without any problem using
fluent-plugin-prometheus v1.0.1. So the article does not need to be updated
beyond this trivial fix.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>