Add a new guide that explains the very basics of Fluentd.
Simizu Taku pointed out that the current Fluentd documentation is
a bit tough for new users in the sense that it lacks overview docs
that explain users the big picture of how Fluentd works.

This article is my attempt to address the problem. This article is
supposed to  1) explain the very basics of Fluentd concisely and
2) navigate users according to their interests.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>