in_monitor_agent: Add a missing documentation for `include_reply`
This patch documents the monitor agent option which has not been
described in the manual page properly.

Note that, as opposed to the name, this option is intended to allow
users to _exclude_ the 'retry' bit from the response, so we would be
better to explain our intention clearly here.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>