rpc: Clean up the documentation
Basically the RPC feature is a thin wrapper over signal handlers
(indeed, these RPC endpoints send corresponding signals on each
request) and it is meant to provide users to more approachable way
to kick those triggers.

This patch polishes up the explanation based on this observation.

TODO: We should explain some endpoints are not usable on Windows.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>