Add an explanation for the "Phi accrual failure detector".
This feature is indeed a quite old beast. It had been introduced in
v0.10.4 (released in 2011) and is being enabled by default in 2018.
Still, there are no proper documentation on how this feature works
and (especially) what the threshold variable `φ` means.

This patch is my attempt to add a general explanation for this failure
detection feature.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>