filter-grep: Explain how to use a pattern with a leading slash
Right now, Fluentd assumes that a match pattern with a leading slash
is written in the regex notation (something like "/foo/").

This can lead to a confusing result if a user naively sets a file path
as a pattern. For example, the following configuration will be parsed
to `RegExp(//)`, not `RegExp(/var/log/)`:

    <regexp>
      key "path"
      pattern /home
    </regexp>

This patch describes how to specify such patterns properly.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>