parser_json: Document new stream_buffer_size param
See fluent/fluentd#2381

In addition to documenting the parameter, update the `in_exec` and
`out_exec_filter` articles to provide a workaround for users
encountering this problem.
This note exists in the v0.12 version of the `in_exec` page but was
absent in the v1.0 docs, so add the whole note.

Closes #624

Signed-off-by: Ben Wheatley <contact@benwh.com>